refactoring ownersList column, move it to separate component to decrease complexity of ExpandedTx component
